refactor: reorganize query builder utilities and moved helper functions
